Add files via upload
This python script is a quick way to list all the files and folders in a user specified directory. 

You just simply use the 'Browse' button to point it to the directory you want to see its contents and hit 'Run'. 
This will generate a text file named "directory_content.txt" in the selected directory, that will contain every filename and its associated extension there.

<b>NOTE</b>: <i>This will NOT include files in sub-directories, but it will include folder names</i>